requirements.txt
# requirements.txt

certifi==2024.8.30
charset-normalizer==3.3.2
i18nice==0.15.5
idna==3.10
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.3
